separate snake utils & constants
