+++ app/src/components/Snake/index.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { SafeAreaView, StyleSheet, View } from "react-native";
 import { Point } from "../../lib/snake";
 import { snakeGreen } from "../../constants/colors";
-import { calculateTopAndLeft } from "../../constants/snake";
+import { calculateTopAndLeft } from "../../utils/snake";
 
 interface Props {
   snake: Point[];
